Add shift quantity column to column folder
