+++ packages/common/src/ui/layout/tables/columns/index.ts
@@ -1,3 +1,4 @@
+export { getEditableQuantityColumn } from './EditableQuantityColumn';
 export { getCheckboxSelectionColumn } from './CheckboxSelectionColumn';
 export { getNameAndColorColumn } from './NameAndColorColumn';
 export * from './types';
