+++ packages/transactions/src/OutboundShipment/DetailView/tabs/GeneralTab.tsx
@@ -1,91 +1,17 @@
 import {
-  Tooltip,
-  Typography,
-  ColumnDefinition,
   ColumnSetBuilder,
   Item,
   RemoteDataTable,
   useColumns,
   useQueryParams,
   useSortedData,
-  TextField,
-  InputAdornment,
-  DomainObject,
+  getEditableQuantityColumn,
 } from '@openmsupply-client/common';
-import React, { FC, useState, useEffect } from 'react';
+import React, { FC } from 'react';
 
 interface GeneralTabProps<T> {
   data: T[];
 }
-
-interface SomeQuantityEntity extends DomainObject {
-  quantity: number;
-  setQuantity: (rowKey: number, newQuantity: number) => void;
-}
-
-const getEditableQuantityColumn = <
-  T extends SomeQuantityEntity
->(): ColumnDefinition<T> => ({
-  key: 'quantity',
-  width: 300,
-  Cell: ({ rowData, rowKey }) => {
-    const { quantity } = rowData;
-    const [buffer, setBuffer] = useState(quantity);
-    const [value, setValue] = useState(quantity);
-    const [error, setError] = useState(false);
-
-    const tryUpdateValue = (
-      event: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
-    ) => {
-      const {
-        target: { value },
-      } = event;
-
-      const asNumber = Number(value);
-      const isValid = Number.isInteger(asNumber) && asNumber >= 0;
-
-      if (isValid) {
-        setValue(asNumber);
-        setError(false);
-      } else {
-        setError(true);
-      }
-
-      setBuffer(asNumber);
-    };
-
-    useEffect(() => {
-      setValue(quantity);
-      setBuffer(quantity);
-      setError(false);
-    }, [rowData]);
-
-    return (
-      <TextField
-        sx={{ maxHeight: 40 }}
-        error={error}
-        variant="filled"
-        size="small"
-        helperText="Incorrect value"
-        hiddenLabel
-        value={buffer}
-        onBlur={() => rowData.setQuantity(Number(rowKey), value)}
-        InputProps={{
-          endAdornment: error ? (
-            <InputAdornment position="end">
-              <Tooltip title="Mate, what you doing?">
-                <Typography sx={{ color: 'red' }}>⚠</Typography>
-              </Tooltip>
-            </InputAdornment>
-          ) : null,
-        }}
-        onChange={tryUpdateValue}
-      />
-    );
-  },
-  label: 'label.quantity',
-  accessor: (row: T) => String(row.quantity),
-});
 
 export const GeneralTab: FC<GeneralTabProps<Item>> = ({ data }) => {
   const { pagination } = useQueryParams({ key: 'quantity' });
